requirements.txt
asgiref==3.9.1
Django==5.2.6
djangorestframework==3.16.1
Pillow==11.0.0
setuptools==80.9.0
sqlparse==0.5.3
wheel==0.45.1
